requirements.txt
flask
PyMySQL
python-dotenv
pathlib
